feat: add ts to backend
