+++ README.md
@@ -2,6 +2,6 @@
 
 ## Local Developement
 In order to get started, run the following commands:
-`cd backend && npm run dev:start`
+`cd backend && yarn dev:start`
 In a new terminal:
 `cd frontend && yarn start`
