+++ backend/template.yaml
@@ -98,6 +98,14 @@ Resources:
           Properties:
             Path: /api/user # Ensure api is proxied with /api/*
             Method: get
+    Metadata:
+      BuildMethod: esbuild
+      BuildProperties:
+        Minify: true
+        Target: "es2020"
+        Sourcemap: true,
+        EntryPoints:
+        - app.ts
 
 Outputs:
   # ServerlessRestApi is an implicit API created out of Events key under Serverless::Function
