+++ frontend/src/redux/reducers/userSlice.ts
@@ -1,5 +1,5 @@
 import { createSlice } from '@reduxjs/toolkit';
-import { User } from '../../types/user';
+import { User } from 'amazonian-prime-types';
 
 export interface UserState {
   value: User | null;
