+++ types/lib/index.d.ts
@@ -1,0 +1,1 @@
+export { User } from "./user";
