+++ types/lib/user.d.ts
@@ -1,0 +1,10 @@
+export interface User {
+    firstName: string;
+    lastName: string;
+    address: {
+        city: string;
+        province: string;
+        streetAddress: string;
+        postalCode: string;
+    };
+}
